feat(`services`): implement signout functionality revoking token
