--- src/app/models/user.ts
@@ -1,11 +1,0 @@
-import { userService } from '../services/user';
-
-async function getInfo(accessToken: string) {
-  const userInfo = await userService.getInfo(accessToken);
-
-  return userInfo;
-}
-
-export const user = Object.freeze({
-  getInfo,
-});
